requirements.txt
Flask
Flask-Cors
Flask-JWT-Extended
python-dotenv
google-auth
gunicorn